Chore: Change Project Structure
