--- readme.md
@@ -1,22 +1,0 @@
-# cnu_announcement_discordBot
-
-## skills
-<img src="https://img.shields.io/badge/dotenv-ECD53F?style=for-the-badge&logo=dotenv&logoColor=white"> <img src="https://img.shields.io/badge/discord-5865F2?style=for-the-badge&logo=discord&logoColor=white"> <img src="https://img.shields.io/badge/poetry-60A5FA?style=for-the-badge&logo=poetry&logoColor=white"> 
-
-## 사용법
-
-You can install them using the following command:
-
-### setting
-```
-poetry install
-```
-and
-```
-base.env를 .env로 바꾸고 안에 값을 채워야됨
-```
-
-### run
-```
-python3 cnu_discord.py
-```
