+++ src/cnudicobot/main.py
@@ -1,0 +1,1 @@
+# import entity here
